add toggle switch for light/dark mode
received some requests for setting dark mode on one's computer, but not in the help center.

created a persistent toggle switch in the header that will allow a user to define the theme mode regardless of what the OS says.

required some tweaks to the logo files, as the `picture` element was not set up to respond to changes in the `styles-light/dark.css` files.