change way the sidebar items are indented
From the comment:

Indent the headings in the auto-generated Table of Contents not by the heading element they're assigned in the article editor, but rather by their relation to the previous heading element.

This way, if an author places an h4 element right after an h1 element, the h4 element is not overly indented.